requirements.txt
Django>=3.1, <3.2
djangorestframework>=3.12, <3.13
django[argon2]
celery>=5.0, <6.0
requests>=2.25, <3.0
PyYAML>=5.0, <6.0
uritemplate>=3.0.1, <4.0